+++ src/client/App.tsx
@@ -11,6 +11,7 @@ import ForgotPasswordScreen from "./screens/ForgotPasswordScreen";
 import HomeScreen from "./screens/HomeScreen";
 import LoginScreen from "./screens/LoginScreen";
 import OrganizationScreen from "./screens/OrganizationScreen";
+import OrganizationAdminScreen from "./screens/OrganizationAdminScreen";
 import ProjectScreen from "./screens/ProjectScreen";
 import RegistrationScreen from "./screens/RegistrationScreen";
 import ResetPasswordScreen from "./screens/ResetPasswordScreen";
@@ -45,6 +46,9 @@ const App = () => (
           <HomeScreen />
         </PrivateRoute>
         <Route path="/o/:organizationSlug" exact={true} component={OrganizationScreen} />
+        <PrivateRoute path="/o/:organizationSlug/admin" exact={true}>
+          <OrganizationAdminScreen />
+        </PrivateRoute>
         <Route path="/projects/:projectId" exact={true} component={ProjectScreen} />
         <Route path="/login" exact={true} component={LoginScreen} />
         <Route path="/register" exact={true} component={RegistrationScreen} />