+++ src/client/actions.ts
@@ -4,6 +4,7 @@ import * as authActions from "./actions/auth";
 import * as districtDrawingActions from "./actions/districtDrawing";
 import * as organizationActions from "./actions/organization";
 import * as organizationJoinActions from "./actions/organizationJoin";
+import * as organizationProjectActions from "./actions/organizationProjects";
 import * as projectDataActions from "./actions/projectData";
 import * as projectsActions from "./actions/projects";
 import * as regionConfigActions from "./actions/regionConfig";
@@ -14,6 +15,7 @@ export type AuthAction = ActionType<typeof authActions>;
 export type DistrictDrawingAction = ActionType<typeof districtDrawingActions>;
 export type OrganizationAction = ActionType<typeof organizationActions>;
 export type OrganizationJoinAction = ActionType<typeof organizationJoinActions>;
+export type OrganizationProjectsAction = ActionType<typeof organizationProjectActions>;
 export type ProjectDataAction = ActionType<typeof projectDataActions>;
 export type ProjectsAction = ActionType<typeof projectsActions>;
 export type RegionConfigAction = ActionType<typeof regionConfigActions>;
@@ -24,6 +26,7 @@ export type Action =
   | AuthAction
   | DistrictDrawingAction
   | OrganizationAction
+  | OrganizationProjectsAction
   | OrganizationJoinAction
   | ProjectDataAction
   | ProjectsAction