+++ src/manage/src/lib/dbUtils.ts
@@ -4,6 +4,7 @@ import { Organization } from "../../../server/src/organizations/entities/organiz
 import { ProjectTemplate } from "../../../server/src/project-templates/entities/project-template.entity";
 import { RegionConfig } from "../../../server/src/region-configs/entities/region-config.entity";
 import { User } from "../../../server/src/users/entities/user.entity";
+import { Project } from "../../../server/src/projects/entities/project.entity";
 
 export const connectionOptions: ConnectionOptions = {
   type: "postgres",
@@ -12,7 +13,7 @@ export const connectionOptions: ConnectionOptions = {
   username: process.env.POSTGRES_USER,
   password: process.env.POSTGRES_PASSWORD,
   database: process.env.POSTGRES_DB,
-  entities: [Chamber, Organization, ProjectTemplate, RegionConfig, User],
+  entities: [Chamber, Organization, ProjectTemplate, RegionConfig, User, Project],
   logging: true,
   synchronize: false
 };