+++ src/server/src/organizations/entities/organization.entity.ts
@@ -1,4 +1,13 @@
-import { Column, Entity, JoinTable, OneToMany, PrimaryGeneratedColumn, ManyToMany } from "typeorm";
+import {
+  Column,
+  Entity,
+  JoinTable,
+  JoinColumn,
+  OneToMany,
+  PrimaryGeneratedColumn,
+  ManyToMany,
+  ManyToOne
+} from "typeorm";
 
 import { IOrganization } from "../../../../shared/entities";
 import { User } from "../../users/entities/user.entity";
@@ -39,6 +48,10 @@ export class Organization implements IOrganization {
   )
   @JoinTable()
   users: User[];
+
+  @ManyToOne(() => User, { nullable: true })
+  @JoinColumn({ name: "user_id" })
+  admin: User;
 
   @OneToMany(
     () => ProjectTemplate,