+++ src/server/src/project-templates/entities/project-template.entity.ts
@@ -1,12 +1,13 @@
-import { Column, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from "typeorm";
+import { Column, Entity, JoinColumn, ManyToOne, OneToMany, PrimaryGeneratedColumn } from "typeorm";
 
-import { DistrictsDefinition, IProjectTemplate } from "../../../../shared/entities";
+import { DistrictsDefinition, IProjectTemplateWithProjects } from "../../../../shared/entities";
 import { Chamber } from "../../chambers/entities/chamber.entity";
 import { Organization } from "../../organizations/entities/organization.entity";
 import { RegionConfig } from "../../region-configs/entities/region-config.entity";
+import { Project } from "../../projects/entities/project.entity";
 
 @Entity()
-export class ProjectTemplate implements IProjectTemplate {
+export class ProjectTemplate implements IProjectTemplateWithProjects {
   @PrimaryGeneratedColumn("uuid")
   id: string;
 
@@ -17,13 +18,23 @@ export class ProjectTemplate implements IProjectTemplate {
   @Column({ type: "character varying" })
   name: string;
 
-  @ManyToOne(() => RegionConfig, { nullable: false })
+  @ManyToOne(
+    () => RegionConfig,
+    regionConfig => regionConfig.projectTemplates,
+    { nullable: false }
+  )
   @JoinColumn({ name: "region_config_id" })
   regionConfig: RegionConfig;
 
   @ManyToOne(() => Chamber, { nullable: true })
   @JoinColumn({ name: "chamber_id" })
   chamber?: Chamber;
+
+  @OneToMany(
+    type => Project,
+    project => project.projectTemplate
+  )
+  projects: Project[];
 
   @Column({ name: "number_of_districts", type: "integer" })
   numberOfDistricts: number;